refactor: architectural enhancements
