+++ src/app/factory.py
@@ -2,12 +2,12 @@ from flask import Flask
 from flasgger import Swagger
 
 from src.app.schemas import ma
-from src.app.infra.logging import logging_infra
+from src.infra.logging import logging_infra
 from src.app.configs.environ import DefaultConfig
 from src.app.utils.helpers.logging import get_logger
-from src.app.infra.middlewares.errors import register_error_handlers
-from src.app.infra.signals import register_gunicorn_signal_handler
-from src.app.infra.middlewares.user_session import trigger_user_session_service
+from src.app.middlewares.errors import register_error_handlers
+from src.infra.signals import register_gunicorn_signal_handler
+from src.app.middlewares.user_session import trigger_user_session_service
 
 log = get_logger(__name__)
