+++ src/app/middlewares/errors.py
@@ -3,7 +3,7 @@ from flask import json, Flask, Response as FlaskResponse
 from werkzeug.exceptions import default_exceptions, HTTPException, InternalServerError
 
 from src.app.utils.helpers.logging import get_logger
-from src.app.infra.exceptions import RedisUnavailableError
+from src.app.middlewares.http_exceptions import RedisUnavailableError
 from src.app.utils.helpers.json import make_error_response
 
 log = get_logger(__name__)
