+++ src/app/middlewares/http_exceptions.py
