+++ src/app/services/oath/totp.py
@@ -1,7 +1,7 @@
 from pyotp import random_base32
 from src.app.services.oath import BaseOTPService
 from src.app.utils.helpers.logging import get_logger
-from src.app.infra.aes_cipher import aes_cipher_infra
+from src.infra.aes_cipher import aes_cipher_infra
 from src.core.entities.totp_entity import TOTPEntity
 from src.core.services.totp_generator import TOTPGenerator
