+++ src/app/utils/helpers/yaml.py
@@ -11,6 +11,6 @@ def get_project_version(dir_path: str) -> str:
             data = load(file, Loader=SafeLoader)
             return data["version"]
     except Exception as e:
-        from src.app.infra.signals import terminate_server
+        from src.infra.signals import terminate_server
         log.error(f"Error loading version.yaml: {e}")
         terminate_server()
