+++ src/tests/unit/app/middlewares/test_auth.py
@@ -4,7 +4,7 @@ from _pytest.python_api import raises
 from werkzeug.exceptions import Unauthorized
 
 from src.tests.helpers import basic_auth
-from src.app.infra.middlewares.auth import auth_middleware
+from src.app.middlewares.auth import auth_middleware
 
 
 @auth_middleware
